+++ src/srclist/ao/write.rs
@@ -107,7 +107,7 @@ pub(crate) fn write_source_list<T: std::io::Write>(
                     writeln!(buf, "    }}")?;
                 }
 
-                FluxDensityType::List { fds } => {
+                FluxDensityType::List(fds) => {
                     for fd in fds {
                         writeln!(buf, "    measurement {{")?;
                         writeln!(buf, "      frequency {} MHz", fd.freq / 1e6)?;